Reorganized to make unit testing easier
